+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added option to enable that clicking on a player name in chat opens the profile viewer of NotEnoughUpdates (to fix SkyHanni breaking the default NEU feature).
 - Added support for new SBA chat icon feature (show profile type and faction in chat)
 - Added hide the damage, ability damage and defence orbs that spawn when the healer is killing mobs
+- Added hide the golden fairy that follows the healer in dungeon.
 
 ### Fixed
 - Links in player chat are clickable again